Auth에 컨트롤러와 서비스 계층이 분리되지 않아서 부적절한 정보들이 컨트롤러 계층에 위치해 있던 문제 해결
- 리뷰사항 https://github.com/codesquad-members-2021/airbnb/pull/53#discussion_r642483047
- 코드 클리닝을 진행하고 서비스계층과 컨트롤러 계층의 역할을 구분함
