--- BE/airbnb/src/main/java/com/example/airbnb/auth/GithubAccountInfoResponse.java
@@ -1,36 +1,0 @@
-package com.example.airbnb.auth;
-
-import com.fasterxml.jackson.annotation.JsonSetter;
-
-public class GithubAccountInfoResponse {
-    private String login;
-    private String id;
-    private String avatar_url;
-
-    public String getLogin() {
-        return login;
-    }
-
-    @JsonSetter("login")
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    @JsonSetter("id")
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getAvatar_url() {
-        return avatar_url;
-    }
-
-    @JsonSetter("avatar_url")
-    public void setAvatar_url(String avatar_url) {
-        this.avatar_url = avatar_url;
-    }
-}
